+++ R/crew_controller_group.R
@@ -1,7 +1,6 @@
 #' @title Create a controller group.
 #' @export
-#' @keywords internal
-#' @family controllers
+#' @family user
 #' @description Create an `R6` object to submit tasks and launch workers
 #'   through multiple `crew` controllers.
 #' @param ... `R6` controller objects or lists of `R6` controller objects.
@@ -31,7 +30,7 @@ crew_controller_group <- function(...) {
 
 #' @title Controller group class
 #' @export
-#' @family controllers
+#' @family class
 #' @description `R6` class for controller groups.
 #' @details See [crew_controller_group()].
 #' @examples