+++ tests/testthat/test-crew_controller_local.R
@@ -145,6 +145,17 @@ crew_test("crew_controller_local() substitute = FALSE and quick push", {
   # regular push
   x$push(command = command, substitute = FALSE, name = "substitute")
   x$wait(seconds_timeout = 10)
+  # just list
+  out <- x$schedule$list()[[1L]]
+  expect_equal(out$result[[1L]], 5L)
+  expect_equal(out$name, "substitute")
+  expect_true(is.numeric(out$seconds))
+  expect_false(anyNA(out$seconds))
+  expect_true(out$seconds >= 0)
+  expect_true(anyNA(out$error))
+  expect_true(anyNA(out$warnings))
+  expect_true(anyNA(out$trace))
+  # full pop
   out <- x$pop(scale = FALSE)
   expect_equal(out$result[[1]], 5L)
   expect_equal(out$name, "substitute")
@@ -155,7 +166,7 @@ crew_test("crew_controller_local() substitute = FALSE and quick push", {
   expect_true(anyNA(out$warnings))
   expect_true(anyNA(out$trace))
   # quick push
-  x$quick_push(command = command, name = "substitute")
+  x$shove(command = command, name = "substitute")
   x$wait(seconds_timeout = 10)
   out <- x$pop(scale = FALSE)
   expect_equal(out$result[[1]], 5L)